Replacing Enduro with the new implementation
